+++ eds-aic-20230930/core/Models/DescribeAndroidInstancesRequest.cs
@@ -136,6 +136,10 @@ namespace AlibabaCloud.SDK.Eds_aic20230930.Models
         [Validation(Required=false)]
         public List<string> OfficeSiteIds { get; set; }
 
+        [NameInMap("QosRuleIds")]
+        [Validation(Required=false)]
+        public List<string> QosRuleIds { get; set; }
+
         /// <summary>
         /// <para>The sales mode.</para>
         /// <para>Valid values:</para>