+++ eds-aic-20230930/core/Models/DescribeDisplayConfigRequest.cs
@@ -1,0 +1,18 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Eds_aic20230930.Models
+{
+    public class DescribeDisplayConfigRequest : TeaModel {
+        [NameInMap("AndroidInstanceIds")]
+        [Validation(Required=false)]
+        public List<string> AndroidInstanceIds { get; set; }
+
+    }
+
+}