+++ eds-aic-20230930/core/Models/ModifyDisplayConfigRequest.cs
@@ -1,0 +1,64 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Eds_aic20230930.Models
+{
+    public class ModifyDisplayConfigRequest : TeaModel {
+        [NameInMap("AndroidInstanceIds")]
+        [Validation(Required=false)]
+        public List<string> AndroidInstanceIds { get; set; }
+
+        [NameInMap("DisplayConfig")]
+        [Validation(Required=false)]
+        public ModifyDisplayConfigRequestDisplayConfig DisplayConfig { get; set; }
+        public class ModifyDisplayConfigRequestDisplayConfig : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>240</para>
+            /// </summary>
+            [NameInMap("Dpi")]
+            [Validation(Required=false)]
+            public int? Dpi { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>null</para>
+            /// </summary>
+            [NameInMap("Fps")]
+            [Validation(Required=false)]
+            public int? Fps { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>off</para>
+            /// </summary>
+            [NameInMap("LockResolution")]
+            [Validation(Required=false)]
+            public string LockResolution { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>1920</para>
+            /// </summary>
+            [NameInMap("ResolutionHeight")]
+            [Validation(Required=false)]
+            public int? ResolutionHeight { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>720</para>
+            /// </summary>
+            [NameInMap("ResolutionWidth")]
+            [Validation(Required=false)]
+            public int? ResolutionWidth { get; set; }
+
+        }
+
+    }
+
+}