+++ eds-aic-20230930/ChangeLog.md
@@ -1,3 +1,10 @@
+2025-05-14 Version: 4.6.0
+- Support API DescribeDisplayConfig.
+- Support API ModifyDisplayConfig.
+- Update API DescribeAndroidInstances: add request parameters QosRuleIds.
+- Update API DescribeAndroidInstances: add response parameters Body.InstanceModel.$.QosRuleId.
+
+
 2025-04-27 Version: 4.5.1
 - Generated csharp 2023-09-30 for eds-aic.